www/requirements.txt
Django>=3.2,<4.0
gunicorn>=20.1.0,<21.0.0
psycopg2-binary>=2.9.1,<3.0.0
djangorestframework>=3.14.0,<4.0.0
django-cors-headers>=3.13.0,<4.0.0